Make compatibility package more compatible
