+++ server/src/test/java/com/vaadin/tests/server/ClassesSerializableTest.java
@@ -54,6 +54,9 @@ public class ClassesSerializableTest {
             "com\\.vaadin\\.ui\\.themes\\..*", //
             // exact class level filtering
             "com\\.vaadin\\.event\\.FieldEvents", //
+            "com\\.vaadin\\.util\\.FileTypeResolver",
+            "com\\.vaadin\\.util\\.FileTypeResolver",
+
             "com\\.vaadin\\.event\\.LayoutEvents", //
             "com\\.vaadin\\.event\\.MouseEvents", //
             "com\\.vaadin\\.event\\.UIEvents", //
